Add `control_flow` module for `for_loop` and `while_loop` (#7017)
**Context:**

`for_loop` and `while_loop` are currently located in `qjit_api.py`. This
isn't an obvious location for these functions now that they have wider
use and are "core" pennylane features.

**Description of the Change:**

Adds a `control_flow` module and places `for_loop` and `while_loop` into
it.

**Benefits:**

Easier to find control flow code.

**Possible Drawbacks:**

Should consider moving `qml.cond` too, but it depends on `Operator`,
queuing, `MeasurementValue`, etc.

**Related GitHub Issues:**

[sc-85453]

---------
