+++ pennylane/__init__.py
@@ -21,7 +21,11 @@ from pennylane.boolean_fn import BooleanFn
 import pennylane.numpy
 from pennylane.queuing import QueuingManager, apply
 
+import pennylane.compiler
+from pennylane.compiler import qjit
 import pennylane.capture
+import pennylane.control_flow
+from pennylane.control_flow import for_loop, while_loop
 import pennylane.kernels
 import pennylane.math
 import pennylane.operation
@@ -146,9 +150,6 @@ from pennylane.drawer import draw, draw_mpl
 
 # pylint:disable=wrong-import-order
 import pennylane.logging  # pylint:disable=wrong-import-order
-
-from pennylane.compiler import qjit, while_loop, for_loop
-import pennylane.compiler
 
 import pennylane.data
