+++ pennylane/capture/primitives.py
@@ -17,7 +17,8 @@ created in pennylane.
 
 It has a jax dependency and should be located in a standard import path.
 """
-from pennylane.compiler.qjit_api import _get_for_loop_qfunc_prim, _get_while_loop_qfunc_prim
+from pennylane.control_flow.for_loop import _get_for_loop_qfunc_prim
+from pennylane.control_flow.while_loop import _get_while_loop_qfunc_prim
 from pennylane.measurements.mid_measure import _create_mid_measure_primitive
 from pennylane.ops.op_math.adjoint import _get_adjoint_qfunc_prim
 from pennylane.ops.op_math.condition import _get_cond_qfunc_prim
