+++ pennylane/compiler/qjit_api.py
@@ -13,19 +13,7 @@
 # limitations under the License.
 """QJIT compatible quantum and compilation operations API"""
 
-import functools
-from collections.abc import Callable
-
-import pennylane as qml
-from pennylane.capture.flatfn import FlatFn
-
-from .compiler import (
-    AvailableCompilers,
-    CompileError,
-    _check_compiler_version,
-    active_compiler,
-    available,
-)
+from .compiler import AvailableCompilers, CompileError, _check_compiler_version, available
 
 
 def qjit(fn=None, *args, compiler="catalyst", **kwargs):  # pylint:disable=keyword-arg-before-vararg
@@ -305,444 +293,3 @@ def qjit(fn=None, *args, compiler="catalyst", **kwargs):  # pylint:disable=keywo
     compilers = AvailableCompilers.names_entrypoints
     qjit_loader = compilers[compiler]["qjit"].load()
     return qjit_loader(fn=fn, *args, **kwargs)
-
-
-def while_loop(cond_fn):
-    """A :func:`~.qjit` compatible for-loop for PennyLane programs. When
-    used without :func:`~.qjit`, this function will fall back to a standard
-    Python for loop.
-
-    This decorator provides a functional version of the traditional while loop,
-    similar to `jax.lax.while_loop <https://jax.readthedocs.io/en/latest/_autosummary/jax.lax.while_loop.html>`__.
-    That is, any variables that are modified across iterations need to be provided as
-    inputs and outputs to the loop body function:
-
-    - Input arguments contain the value of a variable at the start of an
-      iteration
-
-    - Output arguments contain the value at the end of the iteration. The
-      outputs are then fed back as inputs to the next iteration.
-
-    The final iteration values are also returned from the transformed function.
-
-    The semantics of ``while_loop`` are given by the following Python pseudocode:
-
-    .. code-block:: python
-
-        def while_loop(cond_fn, body_fn, *args):
-            while cond_fn(*args):
-                args = body_fn(*args)
-            return args
-
-    Args:
-        cond_fn (Callable): the condition function in the while loop
-
-    Returns:
-        Callable: A wrapper around the while-loop function.
-
-    Raises:
-        CompileError: if the compiler is not installed
-
-    .. seealso:: :func:`~.for_loop`, :func:`~.qjit`
-
-    **Example**
-
-    .. code-block:: python
-
-        dev = qml.device("lightning.qubit", wires=1)
-
-        @qml.qnode(dev)
-        def circuit(x: float):
-
-            @qml.while_loop(lambda x: x < 2.0)
-            def loop_rx(x):
-                # perform some work and update (some of) the arguments
-                qml.RX(x, wires=0)
-                return x ** 2
-
-            # apply the while loop
-            loop_rx(x)
-
-            return qml.expval(qml.Z(0))
-
-    >>> circuit(1.6)
-    -0.02919952
-
-    ``while_loop`` is also :func:`~.qjit` compatible; when used with the
-    :func:`~.qjit` decorator, the while loop will not be unrolled, and instead
-    will be captured as-is during compilation and executed during runtime:
-
-    >>> qml.qjit(circuit)(1.6)
-    Array(-0.02919952, dtype=float64)
-    """
-
-    if active_jit := active_compiler():
-        compilers = AvailableCompilers.names_entrypoints
-        ops_loader = compilers[active_jit]["ops"].load()
-        return ops_loader.while_loop(cond_fn)
-
-    # if there is no active compiler, simply interpret the while loop
-    # via the Python interpretor.
-    def _decorator(body_fn: Callable) -> Callable:
-        """Transform that will call the input ``body_fn`` until the closure variable ``cond_fn`` is met.
-
-        Args:
-            body_fn (Callable):
-
-        Closure Variables:
-            cond_fn (Callable):
-
-        Returns:
-            Callable: a callable with the same signature as ``body_fn`` and ``cond_fn``.
-        """
-        return WhileLoopCallable(cond_fn, body_fn)
-
-    return _decorator
-
-
-@functools.lru_cache
-def _get_while_loop_qfunc_prim():
-    """Get the while_loop primitive for quantum functions."""
-
-    # pylint: disable=import-outside-toplevel
-    import jax
-
-    from pennylane.capture.custom_primitives import NonInterpPrimitive
-
-    while_loop_prim = NonInterpPrimitive("while_loop")
-    while_loop_prim.multiple_results = True
-    while_loop_prim.prim_type = "higher_order"
-
-    # pylint: disable=too-many-arguments
-    @while_loop_prim.def_impl
-    def _(
-        *args,
-        jaxpr_body_fn,
-        jaxpr_cond_fn,
-        body_slice,
-        cond_slice,
-        args_slice,
-        abstract_shapes_slice,
-    ):
-
-        jaxpr_consts_body = args[body_slice]
-        jaxpr_consts_cond = args[cond_slice]
-        init_state = args[args_slice]
-        abstract_shapes = args[abstract_shapes_slice]
-        # If cond_fn(*init_state) is False, return the initial state
-        fn_res = init_state
-        while jax.core.eval_jaxpr(jaxpr_cond_fn, jaxpr_consts_cond, *abstract_shapes, *fn_res)[0]:
-            fn_res = jax.core.eval_jaxpr(
-                jaxpr_body_fn, jaxpr_consts_body, *abstract_shapes, *fn_res
-            )
-
-        return fn_res
-
-    @while_loop_prim.def_abstract_eval
-    def _(*args, args_slice, **__):
-        return args[args_slice]
-
-    return while_loop_prim
-
-
-class WhileLoopCallable:  # pylint:disable=too-few-public-methods
-    """Base class to represent a while loop. This class
-    when called with an initial state will execute the while
-    loop via the Python interpreter.
-
-    Args:
-        cond_fn (Callable): the condition function in the while loop
-        body_fn (Callable): the function that is executed within the while loop
-    """
-
-    def __init__(self, cond_fn, body_fn):
-        self.cond_fn = cond_fn
-        self.body_fn = body_fn
-
-    def _call_capture_disabled(self, *init_state):
-        args = init_state
-        fn_res = args if len(args) > 1 else args[0] if len(args) == 1 else None
-
-        while self.cond_fn(*args):
-            fn_res = self.body_fn(*args)
-            args = fn_res if len(args) > 1 else (fn_res,) if len(args) == 1 else ()
-
-        return fn_res
-
-    def _call_capture_enabled(self, *init_state):
-
-        import jax  # pylint: disable=import-outside-toplevel
-
-        while_loop_prim = _get_while_loop_qfunc_prim()
-
-        abstracted_axes, abstract_shapes = qml.capture.determine_abstracted_axes(init_state)
-
-        flat_body_fn = FlatFn(self.body_fn)
-        jaxpr_body_fn = jax.make_jaxpr(flat_body_fn, abstracted_axes=abstracted_axes)(*init_state)
-        jaxpr_cond_fn = jax.make_jaxpr(self.cond_fn, abstracted_axes=abstracted_axes)(*init_state)
-
-        body_consts = slice(0, len(jaxpr_body_fn.consts))
-        cond_consts = slice(body_consts.stop, body_consts.stop + len(jaxpr_cond_fn.consts))
-        abstract_shapes_slice = slice(cond_consts.stop, cond_consts.stop + len(abstract_shapes))
-        args_slice = slice(abstract_shapes_slice.stop, None)
-
-        flat_args, _ = jax.tree_util.tree_flatten(init_state)
-        results = while_loop_prim.bind(
-            *jaxpr_body_fn.consts,
-            *jaxpr_cond_fn.consts,
-            *abstract_shapes,
-            *flat_args,
-            jaxpr_body_fn=jaxpr_body_fn.jaxpr,
-            jaxpr_cond_fn=jaxpr_cond_fn.jaxpr,
-            body_slice=body_consts,
-            cond_slice=cond_consts,
-            args_slice=args_slice,
-            abstract_shapes_slice=abstract_shapes_slice,
-        )
-        assert flat_body_fn.out_tree is not None, "Should be set when constructing the jaxpr"
-        return jax.tree_util.tree_unflatten(flat_body_fn.out_tree, results)
-
-    def __call__(self, *init_state):
-
-        if qml.capture.enabled():
-            return self._call_capture_enabled(*init_state)
-
-        return self._call_capture_disabled(*init_state)
-
-
-def for_loop(start, stop=None, step=1):
-    """for_loop([start, ]stop[, step])
-    A :func:`~.qjit` compatible for-loop for PennyLane programs. When
-    used without :func:`~.qjit`, this function will fall back to a standard
-    Python for loop.
-
-    This decorator provides a functional version of the traditional
-    for-loop, similar to `jax.cond.fori_loop <https://jax.readthedocs.io/en/latest/_autosummary/jax.lax.fori_loop.html>`__.
-    That is, any variables that are modified across iterations need to be provided
-    as inputs/outputs to the loop body function:
-
-    - Input arguments contain the value of a variable at the start of an
-      iteration.
-
-    - output arguments contain the value at the end of the iteration. The
-      outputs are then fed back as inputs to the next iteration.
-
-    The final iteration values are also returned from the transformed
-    function.
-
-    The semantics of ``for_loop`` are given by the following Python pseudo-code:
-
-    .. code-block:: python
-
-        def for_loop(start, stop, step, loop_fn, *args):
-            for i in range(start, stop, step):
-                args = loop_fn(i, *args)
-            return args
-
-    Unlike ``jax.cond.fori_loop``, the step can be negative if it is known at tracing time
-    (i.e., constant). If a non-constant negative step is used, the loop will produce no iterations.
-
-    .. note::
-
-        This function can be used in the following different ways:
-
-        1. ``for_loop(stop)``:  Values are generated within the interval ``[0, stop)``
-        2. ``for_loop(start, stop)``: Values are generated within the interval ``[start, stop)``
-        3. ``for_loop(start, stop, step)``: Values are generated within the interval ``[start, stop)``,
-           with spacing between the values given by ``step``
-
-    Args:
-        start (int, optional): starting value of the iteration index.
-            The default start value is ``0``
-        stop (int): upper bound of the iteration index
-        step (int, optional): increment applied to the iteration index at the end of
-            each iteration. The default step size is ``1``
-
-    Returns:
-        Callable[[int, ...], ...]: A wrapper around the loop body function.
-        Note that the loop body function must always have the iteration index as its first
-        argument, which can be used arbitrarily inside the loop body. As the value of the index
-        across iterations is handled automatically by the provided loop bounds, it must not be
-        returned from the function.
-
-    .. seealso:: :func:`~.while_loop`, :func:`~.qjit`
-
-    **Example**
-
-    .. code-block:: python
-
-        dev = qml.device("lightning.qubit", wires=1)
-
-        @qml.qnode(dev)
-        def circuit(n: int, x: float):
-
-            @qml.for_loop(0, n, 1)
-            def loop_rx(i, x):
-                # perform some work and update (some of) the arguments
-                qml.RX(x, wires=0)
-
-                # update the value of x for the next iteration
-                return jnp.sin(x)
-
-            # apply the for loop
-            final_x = loop_rx(x)
-
-            return qml.expval(qml.Z(0))
-
-    >>> circuit(7, 1.6)
-    array(0.97926626)
-
-    ``for_loop`` is also :func:`~.qjit` compatible; when used with the
-    :func:`~.qjit` decorator, the for loop will not be unrolled, and instead
-    will be captured as-is during compilation and executed during runtime:
-
-    >>> qml.qjit(circuit)(7, 1.6)
-    Array(0.97926626, dtype=float64)
-
-    .. note::
-
-        Please see the Catalyst :doc:`quickstart guide <catalyst:dev/quick_start>`,
-        as well as the :doc:`sharp bits and debugging tips <catalyst:dev/sharp_bits>`
-        page for an overview of using quantum just-in-time compilation.
-
-    """
-    if stop is None:
-        start, stop = 0, start
-
-    if active_jit := active_compiler():
-        compilers = AvailableCompilers.names_entrypoints
-        ops_loader = compilers[active_jit]["ops"].load()
-        return ops_loader.for_loop(start, stop, step)
-
-    # if there is no active compiler, simply interpret the for loop
-    # via the Python interpretor.
-    def _decorator(body_fn):
-        """Transform that will call the input ``body_fn`` within a for loop defined by the closure variables start, stop, and step.
-
-        Args:
-            body_fn (Callable): The function called within the for loop. Note that the loop body
-                function must always have the iteration index as its first
-                argument, which can be used arbitrarily inside the loop body. As the value of the index
-                across iterations is handled automatically by the provided loop bounds, it must not be
-                returned from the function.
-
-        Closure Variables:
-            start (int): starting value of the iteration index
-            stop (int): (exclusive) upper bound of the iteration index
-            step (int): increment applied to the iteration index at the end of each iteration
-
-        Returns:
-            Callable: a callable with the same signature as ``body_fn``
-        """
-        return ForLoopCallable(start, stop, step, body_fn)
-
-    return _decorator
-
-
-@functools.lru_cache
-def _get_for_loop_qfunc_prim():
-    """Get the loop_for primitive for quantum functions."""
-
-    # pylint: disable=import-outside-toplevel
-    import jax
-
-    from pennylane.capture.custom_primitives import NonInterpPrimitive
-
-    for_loop_prim = NonInterpPrimitive("for_loop")
-    for_loop_prim.multiple_results = True
-    for_loop_prim.prim_type = "higher_order"
-
-    # pylint: disable=too-many-arguments
-    @for_loop_prim.def_impl
-    def _(start, stop, step, *args, jaxpr_body_fn, consts_slice, args_slice, abstract_shapes_slice):
-
-        consts = args[consts_slice]
-        init_state = args[args_slice]
-        abstract_shapes = args[abstract_shapes_slice]
-
-        # in case start >= stop, return the initial state
-        fn_res = init_state
-
-        for i in range(start, stop, step):
-            fn_res = jax.core.eval_jaxpr(jaxpr_body_fn, consts, *abstract_shapes, i, *fn_res)
-
-        return fn_res
-
-    # pylint: disable=unused-argument
-    @for_loop_prim.def_abstract_eval
-    def _(start, stop, step, *args, args_slice, **_):
-        return args[args_slice]
-
-    return for_loop_prim
-
-
-class ForLoopCallable:  # pylint:disable=too-few-public-methods
-    """Base class to represent a for loop. This class
-    when called with an initial state will execute the while
-    loop via the Python interpreter.
-
-    Args:
-        start (int): starting value of the iteration index
-        stop (int): (exclusive) upper bound of the iteration index
-        step (int): increment applied to the iteration index at the end of each iteration
-        body_fn (Callable): The function called within the for loop. Note that the loop body
-            function must always have the iteration index as its first
-            argument, which can be used arbitrarily inside the loop body. As the value of the index
-            across iterations is handled automatically by the provided loop bounds, it must not be
-            returned from the function.
-    """
-
-    def __init__(self, start, stop, step, body_fn):
-        self.start = start
-        self.stop = stop
-        self.step = step
-        self.body_fn = body_fn
-
-    def _call_capture_disabled(self, *init_state):
-        args = init_state
-        fn_res = args if len(args) > 1 else args[0] if len(args) == 1 else None
-
-        for i in range(self.start, self.stop, self.step):
-            fn_res = self.body_fn(i, *args)
-            args = fn_res if len(args) > 1 else (fn_res,) if len(args) == 1 else ()
-
-        return fn_res
-
-    def _call_capture_enabled(self, *init_state):
-
-        import jax  # pylint: disable=import-outside-toplevel
-
-        for_loop_prim = _get_for_loop_qfunc_prim()
-
-        abstracted_axes, abstract_shapes = qml.capture.determine_abstracted_axes((0, *init_state))
-
-        flat_fn = FlatFn(self.body_fn)
-        jaxpr_body_fn = jax.make_jaxpr(flat_fn, abstracted_axes=abstracted_axes)(0, *init_state)
-
-        consts_slice = slice(0, len(jaxpr_body_fn.consts))
-        abstract_shapes_slice = slice(consts_slice.stop, consts_slice.stop + len(abstract_shapes))
-        args_slice = slice(abstract_shapes_slice.stop, None)
-
-        flat_args, _ = jax.tree_util.tree_flatten(init_state)
-
-        results = for_loop_prim.bind(
-            self.start,
-            self.stop,
-            self.step,
-            *jaxpr_body_fn.consts,
-            *abstract_shapes,
-            *flat_args,
-            jaxpr_body_fn=jaxpr_body_fn.jaxpr,
-            consts_slice=consts_slice,
-            args_slice=args_slice,
-            abstract_shapes_slice=abstract_shapes_slice,
-        )
-        assert flat_fn.out_tree is not None
-        return jax.tree_util.tree_unflatten(flat_fn.out_tree, results)
-
-    def __call__(self, *init_state):
-
-        if qml.capture.enabled():
-            return self._call_capture_enabled(*init_state)
-
-        return self._call_capture_disabled(*init_state)
